requirements.txt
arrow==0.13.2
asgiref==3.7.2
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.8.2
boto3==1.26.2
botocore==1.29.132
cachetools==4.2.1
certifi==2018.10.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==37.0.4
dataclasses==0.6
defusedxml==0.6.0
Django==3.2.12
django-cors-middleware==1.3.1
django-dynamic-db-router==0.3.0
django-file-md5==1.0.3
django-filter==2.0.0
django-mssql-backend==2.8.1
django-oauth-toolkit==1.2.0
django-redis==4.10.0
django-storages==1.7.1
django-stubs==0.12.1
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.3
django-wkhtmltopdf==3.2.0
djangorestframework==3.9.0
djangorestframework-camel-case==1.0.3
djangorestframework-stubs==0.2.0
docopt==0.6.2
docutils==0.14
drf-yasg==1.17.0
exponent-server-sdk==0.3.1
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.18.1
idna==2.7
importlib-metadata==6.7.0
inflection==0.3.1
isort==4.3.20
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
lazy-object-proxy==1.4.1
lml==0.0.9
Markdown==3.4.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
num2words==0.5.10
numpy==1.17.3
oauthlib==3.1.0
odfpy==1.4.0
packaging==20.1
pandas==0.25.1
Pillow==8.4.0
pjbank==0.9.dev4
protobuf==3.14.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.21
pydocstyle==5.1.1
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-ods==0.5.6
PyJWT==2.4.0
pylint==2.3.1
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyodbc==4.0.26
pyparsing==2.4.6
pypdf==3.16.4
python-dateutil==2.7.5
python-json-logger==2.0.2
pytz==2018.6
PyYAML==5.1.2
redis==3.2.1
requests==2.30.0
requests-oauthlib==1.3.0
rsa==4.7
ruamel.yaml==0.15.97
s3transfer==0.6.1
simplejson==3.16.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==1.9.5
sqlparse==0.3.0
texttable==1.6.2
toml==0.10.1
typed-ast==1.3.4
typing-extensions==4.7.1
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.25.11
user-agents==2.0
watchtower==0.5.5
webencodings==0.5.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.0
xlwt==1.3.0
zipp==3.15.0
